Fix bug with Save
Subclassing BrowserWindow in es6 style seems to cause some strange bugs
in electron. I've changed this so that methods are added to the
individual instances, as that keeps the object as type BrowserWindow.

Not sure what the problem is with BrowserWindow subclasses. If anyone
knows the reason, would love to hear if there's a better solution!